+++ lib/Data/Time/Calendar/CalendarDiffDays.hs
@@ -7,6 +7,7 @@ module Data.Time.Calendar.CalendarDiffDays (
 
 import Control.DeepSeq
 import Data.Data
+import qualified Language.Haskell.TH.Syntax as TH
 
 data CalendarDiffDays = CalendarDiffDays
     { cdMonths :: Integer
@@ -18,6 +19,8 @@ data CalendarDiffDays = CalendarDiffDays
           Data
         , -- | @since 1.9.2
           Typeable
+        , -- | @since 1.13.0
+          TH.Lift
         )
 
 instance NFData CalendarDiffDays where