+++ lib/Data/Time/Calendar/Days.hs
@@ -18,12 +18,13 @@ module Data.Time.Calendar.Days (
 import Control.DeepSeq
 import Data.Data
 import Data.Ix
+import qualified Language.Haskell.TH.Syntax as TH
 
 -- | The Modified Julian Day is a standard count of days, with zero being the day 1858-11-17.
 newtype Day = ModifiedJulianDay
     { toModifiedJulianDay :: Integer
     }
-    deriving (Eq, Ord, Data, Typeable)
+    deriving (Eq, Ord, Data, Typeable, TH.Lift)
 
 instance NFData Day where
     rnf (ModifiedJulianDay a) = rnf a