+++ lib/Data/Time/Calendar/Month.hs
@@ -18,12 +18,13 @@ import Data.Ix
 import Data.Time.Calendar.Days
 import Data.Time.Calendar.Gregorian
 import Data.Time.Calendar.Private
+import qualified Language.Haskell.TH.Syntax as TH
 import Text.ParserCombinators.ReadP
 import Text.Read
 
 -- | An absolute count of common calendar months.
 -- Number is equal to @(year * 12) + (monthOfYear - 1)@.
-newtype Month = MkMonth Integer deriving (Eq, Ord, Data, Typeable)
+newtype Month = MkMonth Integer deriving (Eq, Ord, Data, Typeable, TH.Lift)
 
 instance NFData Month where
     rnf (MkMonth m) = rnf m