+++ lib/Data/Time/Calendar/Week.hs
@@ -16,6 +16,7 @@ import Data.Data
 import Data.Fixed
 import Data.Ix
 import Data.Time.Calendar.Days
+import qualified Language.Haskell.TH.Syntax as TH
 
 data DayOfWeek
     = Monday
@@ -25,7 +26,7 @@ data DayOfWeek
     | Friday
     | Saturday
     | Sunday
-    deriving (Eq, Show, Read, Data, Typeable, Ord, Ix)
+    deriving (Eq, Show, Read, Data, Typeable, Ord, Ix, TH.Lift)
 
 instance NFData DayOfWeek where
     rnf Monday = ()